+++ common/utils.go
@@ -3,6 +3,8 @@ package common
 import (
 	"log"
 	"os"
+	"os/exec"
+	"syscall"
 )
 
 var (
@@ -17,4 +19,21 @@ func FatalOnError(err error) {
 
 func Fatalln(message string) {
 	ErrorLog.Fatalln(message)
+}
+
+func GetExitCode(cmd *exec.Cmd, err error) int {
+	// adapted from https://stackoverflow.com/a/40770011
+	if err != nil {
+		// try to get the exit code
+		if exitError, ok := err.(*exec.ExitError); ok {
+			ws := exitError.Sys().(syscall.WaitStatus)
+			return ws.ExitStatus()
+		}
+		return -1
+
+	} else {
+		// success, exitCode should be 0 if go is ok
+		ws := cmd.ProcessState.Sys().(syscall.WaitStatus)
+		return ws.ExitStatus()
+	}
 }